#465 feat: prod, local ApplicationRunner 초기화 중복 코드 제거
