+++ backend/src/main/java/moheng/runner/data/KeywordDevApplicationRunner.java
@@ -1,4 +1,4 @@
-package moheng.runner.data.prod;
+package moheng.runner.data;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -20,17 +20,17 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@Profile("prod")
+@Profile({"local", "dev", "prod"})
 @Order(3)
 @Component
-public class KeywordProdApplicationRunner implements ApplicationRunner {
+public class KeywordDevApplicationRunner implements ApplicationRunner {
     private final KeywordRepository keywordRepository;
     private final TripRepository tripRepository;
     private final TripKeywordRepository tripKeywordRepository;
 
-    public KeywordProdApplicationRunner(final KeywordRepository keywordRepository,
-                                        final TripRepository tripRepository,
-                                        final TripKeywordRepository tripKeywordRepository) {
+    public KeywordDevApplicationRunner(final KeywordRepository keywordRepository,
+                                       final TripRepository tripRepository,
+                                       final TripKeywordRepository tripKeywordRepository) {
         this.keywordRepository = keywordRepository;
         this.tripRepository = tripRepository;
         this.tripKeywordRepository = tripKeywordRepository;
